# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: EduMaster 360 - School Operating System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── STUDENTS ────────────────────────────────────────────────────────
...
  /classes:
    get:
      operationId: listClasses
      tags: [classes]
      summary: List all classes
      responses:
        "200":
          description: Classes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SchoolClass"
    post:
      operationId: createClass
      tags: [classes]
      summary: Create class
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchoolClass"
  /classes/{id}:
    get:
      operationId: getClass
      tags: [classes]
      summary: Get class by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Class
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchoolClass"

  # ─── MARKS ───────────────────────────────────────────────────────────
...
  /attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: List attendance records
      parameters:
        - name: studentId
          in: query
          schema:
            type: ["integer", "null"]
        - name: classId
          in: query
          schema:
            type: ["integer", "null"]
        - name: date
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"
    post:
      operationId: recordAttendance
      tags: [attendance]
      summary: Record attendance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceInput"
      responses:
        "201":
          description: Recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
...
  /discipline:
    get:
      operationId: listDiscipline
      tags: [discipline]
      summary: List discipline records
      parameters:
        - name: studentId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Discipline records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DisciplineRecord"
    post:
      operationId: createDisciplineRecord
      tags: [discipline]
      summary: Create discipline record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DisciplineInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DisciplineRecord"
  /discipline/{id}:
    patch:
      operationId: updateDisciplineRecord
      tags: [discipline]
      summary: Update discipline record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DisciplineUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DisciplineRecord"

  # ─── ANNOUNCEMENTS ─────────────────────────────────────────────────────
  /announcements:
    get:
      operationId: listAnnouncements
      tags: [announcements]
      summary: List announcements
      parameters:
        - name: visibility
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Announcements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"
    post:
      operationId: createAnnouncement
      tags: [announcements]
      summary: Create announcement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
  /announcements/{id}:
    patch:
      operationId: updateAnnouncement
      tags: [announcements]
      summary: Update announcement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
    delete:
      operationId: deleteAnnouncement
      tags: [announcements]
      summary: Delete announcement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── SUGGESTIONS ────────────────────────────────────────────────────────
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # ─── AUTH ──────────────────────────────────────────────────────────
    LoginInput:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    AuthResult:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, headteacher, dos, teacher, bursar, student, parent]
        avatarUrl:
          type: ["string", "null"]
        linkedId:
          type: ["integer", "null"]
      required: [id, username, name, role]

    # ─── STUDENTS ─────────────────────────────────────────────────────
...
    SchoolClass:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        year:
          type: string
        teacherId:
          type: ["integer", "null"]
        teacherName:
          type: ["string", "null"]
        studentCount:
          type: ["integer", "null"]
        averageScore:
          type: ["number", "null"]
        createdAt:
          type: string
      required: [id, name, year, createdAt]
    ClassInput:
      type: object
      properties:
        name:
          type: string
        year:
          type: string
        teacherId:
          type: ["integer", "null"]
      required: [name, year]

    # ─── MARKS ────────────────────────────────────────────────────────
...
    AttendanceRecord:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: ["string", "null"]
        classId:
          type: ["integer", "null"]
        date:
          type: string
        status:
          type: string
          enum: [present, absent, late, excused]
        reason:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, studentId, date, status, createdAt]
    AttendanceInput:
      type: object
      properties:
        studentId:
          type: integer
        classId:
          type: ["integer", "null"]
        date:
          type: string
        status:
          type: string
          enum: [present, absent, late, excused]
        reason:
          type: ["string", "null"]
      required: [studentId, date, status]
...
    DisciplineRecord:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: ["string", "null"]
        type:
          type: string
          enum: [warning, suspension, commendation, detention, expulsion]
        description:
          type: string
        date:
          type: string
        issuedBy:
          type: ["string", "null"]
        severity:
          type: string
          enum: [low, medium, high]
        resolved:
          type: boolean
        createdAt:
          type: string
      required: [id, studentId, type, description, date, severity, resolved, createdAt]
    DisciplineInput:
      type: object
      properties:
        studentId:
          type: integer
        type:
          type: string
          enum: [warning, suspension, commendation, detention, expulsion]
        description:
          type: string
        date:
          type: string
        severity:
          type: string
          enum: [low, medium, high]
      required: [studentId, type, description, date, severity]
    DisciplineUpdate:
      type: object
      properties:
        resolved:
          type: boolean
        description:
          type: string

    # ─── ANNOUNCEMENTS ────────────────────────────────────────────────
    Announcement:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        authorId:
          type: integer
        authorName:
          type: ["string", "null"]
        visibility:
          type: string
          enum: [public, staff, students, parents]
        category:
          type: string
          enum: [general, event, achievement, alert, academic]
        imageUrl:
          type: ["string", "null"]
        pinned:
          type: boolean
        createdAt:
          type: string
      required: [id, title, content, authorId, visibility, category, pinned, createdAt]
    AnnouncementInput:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        visibility:
          type: string
          enum: [public, staff, students, parents]
        category:
          type: string
          enum: [general, event, achievement, alert, academic]
        imageUrl:
          type: ["string", "null"]
        pinned:
          type: boolean
      required: [title, content, visibility, category]
    AnnouncementUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        visibility:
          type: string
          enum: [public, staff, students, parents]
        category:
          type: string
          enum: [general, event, achievement, alert, academic]
        pinned:
          type: boolean

    # ─── SUGGESTIONS ─────────────────────────────────────────────────
...