# lib/api-zod/src/generated/types/announcementUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EduMaster 360 - School Operating System API
 * OpenAPI spec version: 0.1.0
 */
import type { AnnouncementUpdateCategory } from './announcementUpdateCategory';
import type { AnnouncementUpdateVisibility } from './announcementUpdateVisibility';

export interface AnnouncementUpdate {
  title?: string;
  content?: string;
  visibility?: AnnouncementUpdateVisibility;
  category?: AnnouncementUpdateCategory;
  pinned?: boolean;
}